apiPapel/requirements.txt
asgiref==3.3.1
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.3
idna==2.10
lxml==4.6.3
mysql-client==0.0.1
mysqlclient==2.0.3
parameterized==0.8.1
Pillow==8.2.0
pycparser==2.20
pycryptodome==3.10.1
python-dateutil==2.8.1
pytz==2021.1
reportlab==3.5.67
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
validator==0.7.0
xmltodict==0.12.0
